Refactor: 1. Add ASTDecl 2. Add Visitors (ASTAnalyzer.hpp - StyioASTAnalyzer; CodeGenVisitor - StyioToLLVMIR). Compilation Passed. Next: Separate JIT from StyioToLLVMIR
