+++ src/include/StyioAST/AST.hpp
@@ -3,777 +3,16 @@
 #define STYIO_AST_H_
 
 // [Styio]
-#include "../StyioJIT/StyioJIT_ORC.hpp"
 #include "../StyioToken/Token.hpp"
+#include "../StyioVisitors/ASTAnalyzer.hpp"
+#include "../StyioVisitors/CodeGenVisitor.hpp"
+#include "ASTDecl.hpp"
 
 // [LLVM]
-#include "llvm/Analysis/CGSCCPassManager.h" /* CGSCCAnalysisManager */
-#include "llvm/IR/IRBuilder.h"
-#include "llvm/IR/LLVMContext.h"
-#include "llvm/IR/LegacyPassManager.h" /* FunctionPassManager */
-#include "llvm/IR/Module.h"
-#include "llvm/IR/PassInstrumentation.h" /* PassInstrumentationCallbacks */
-#include "llvm/IR/PassManager.h"         /* LoopAnalysisManager, FunctionAnalysisManager, ModuleAnalysisManager */
 #include "llvm/IR/Value.h"
-#include "llvm/Passes/PassBuilder.h"                 /* PassBuilder */
-#include "llvm/Passes/StandardInstrumentations.h"    /* StandardInstrumentations.h */
-#include "llvm/Support/TargetSelect.h"               /* InitializeNativeTarget, InitializeNativeTargetAsmPrinter, InitializeNativeTargetAsmParser */
-#include "llvm/Transforms/InstCombine/InstCombine.h" /* InstCombinePass */
-#include "llvm/Transforms/Scalar.h"
-#include "llvm/Transforms/Scalar/GVN.h"         /* GVNPass */
-#include "llvm/Transforms/Scalar/Reassociate.h" /* ReassociatePass */
-#include "llvm/Transforms/Scalar/SimplifyCFG.h" /* SimplifyCFGPass */
-#include "llvm/Transforms/Utils.h"
+#include "llvm/IR/Type.h"
 
-using std::string;
-using std::unordered_map;
-using std::vector;
-
-using std::make_shared;
-using std::make_unique;
-using std::shared_ptr;
-using std::unique_ptr;
-
-// Generic Visitor
-template <typename... Types>
-class Visitor;
-
-template <typename T>
-class Visitor<T>
-{
-public:
-  /* Type Inference */
-  virtual void typeInfer(T* t) = 0;
-
-  /* Get LLVM Type */
-  virtual llvm::Type* getLLVMType(T* t) = 0;
-
-  /* LLVM IR Generator */
-  virtual llvm::Value* toLLVMIR(T* t) = 0;
-};
-
-template <typename T, typename... Types>
-class Visitor<T, Types...> : public Visitor<Types...>
-{
-public:
-  using Visitor<Types...>::typeInfer;
-  using Visitor<Types...>::getLLVMType;
-  using Visitor<Types...>::toLLVMIR;
-
-  /* Type Inference */
-  virtual void typeInfer(T* t) = 0;
-
-  /* Get LLVM Type */
-  virtual llvm::Type* getLLVMType(T* t) = 0;
-
-  /* LLVM IR Generator */
-  virtual llvm::Value* toLLVMIR(T* t) = 0;
-};
-
-/* Forward Declaration */
-
-/* Styio AST */
-// Base
-class StyioAST;
-
-// Comment
-class CommentAST;
-
-// None / Empty
-class NoneAST;
-class EmptyAST;
-
-/*
-  Literals:
-  - Bool (Boolean)
-  - Int (Integer)
-  - Float
-  - Char (Character)
-*/
-class BoolAST;
-class IntAST;
-class FloatAST;
-class CharAST;
-
-/*
-  Collections:
-  - string [homo]
-
-  - set [homo]
-  - list [heter]
-
-  - struct [heter]
-  - tuple [heter]
-*/
-class StringAST;
-class SetAST;
-class ListAST;
-
-class StructAST;
-class TupleAST;
-
-/*
-  Variable:
-  - Id
-  - DType (Data Type)
-*/
-class IdAST;
-class DTypeAST;
-
-/*
-  Variable Types:
-  - Var ([Global|Local] Variables)
-  - Arg (Function Arguments)
-  - OptArg (Optional Function Arguments)
-  - OptKwArg (Optional Function Arguments)
-*/
-class VarAST;
-class ArgAST;
-class OptArgAST;
-class OptKwArgAST;
-
-/*
-  Assignment
-  - FlexBind
-  - FinalBind
-*/
-class FlexBindAST;
-class FinalBindAST;
-
-/*
-  Binary Tree:
-  - BinComp (Comparisons)
-  - Cond (Condition)
-  - BinOp (Binary Operations)
-*/
-class BinCompAST;
-class CondAST;
-class BinOpAST;
-
-/*
-  Function
-  - Anonymous Function
-  - Function
-
-  - Call
-*/
-class AnonyFuncAST;
-class FuncAST;
-
-class CallAST;
-
-/*
-  Methods:
-  - SizeOf
-
-  - TypeOf
-  - TypeConvert
-
-  - Range
-  - ListOp
-*/
-class SizeOfAST;
-
-class TypeOfAST;
-class TypeConvertAST;
-
-class ListOpAST;
-
-class RangeAST;
-
-/*
-  Iteration:
-  - Iter (Iterator)
-  - Loop
-*/
-class IterAST;
-class LoopAST;
-
-/*
-  Control Flow
-  - CondFlow (Conditional)
-
-  - End-Of-Line
-  - `pass` ..........
-  - `break` ^^^^^^^^^
-  - `continue` >>>>>>
-  - `return` <<<<<<<<
-*/
-class CondFlowAST;
-
-class EOFAST;
-class PassAST;
-class BreakAST;
-class ContinueAST;
-class ReturnAST;
-
-/*
-  Cases
-  - Cases
-  - MatchCases
-*/
-class CasesAST;
-class MatchCasesAST;
-
-/*
-  Blocks
-  - SideBlock
-  - MainBlock
-*/
-class BlockAST;
-class MainBlockAST;
-
-/*
-  Package Management
-  - ExtPack
-*/
-class ExtPackAST;
-
-/*
-  Conceptions:
-  - Infinite
-*/
-class InfiniteAST;
-
-/*
-  Intermediate Components:
-  - VarTupleAST
-*/
-class VarTupleAST;
-
-class ForwardAST;
-class CheckEqAST;
-class CheckIsinAST;
-class FromToAST;
-
-/*
-  Features:
-  - FmtStr (Format String)
-*/
-class FmtStrAST;
-
-/*
-  Resources:
-  - ResourceAST
-
-  - LocalPath
-  - RemotePath
-  - WebUrl
-  - DBUrl
-*/
-class ResourceAST;
-
-class LocalPathAST;
-class RemotePathAST;
-class WebUrlAST;
-class DBUrlAST;
-
-/*
-  Methods (I/O)
-  - print
-  - read file
-  - write file
-*/
-class PrintAST;
-class ReadFileAST;
-class WriteFileAST;
-
-using StyioVisitor = Visitor<
-  class CommentAST,
-
-  class NoneAST,
-  class EmptyAST,
-
-  class BoolAST,
-  class IntAST,
-  class FloatAST,
-  class CharAST,
-
-  class StringAST,
-  class SetAST,
-  class ListAST,
-
-  class StructAST,
-  class TupleAST,
-
-  class IdAST,
-  class DTypeAST,
-
-  class VarAST,
-  class ArgAST,
-  class OptArgAST,
-  class OptKwArgAST,
-
-  class FlexBindAST,
-  class FinalBindAST,
-
-  class BinCompAST,
-  class CondAST,
-  class BinOpAST,
-
-  class AnonyFuncAST,
-  class FuncAST,
-
-  class CallAST,
-
-  class SizeOfAST,
-  class TypeConvertAST,
-  class ListOpAST,
-  class RangeAST,
-
-  class IterAST,
-  class LoopAST,
-
-  class CondFlowAST,
-
-  class EOFAST,
-  class PassAST,
-  class BreakAST,
-  class ReturnAST,
-
-  class CasesAST,
-  class MatchCasesAST,
-
-  class BlockAST,
-  class MainBlockAST,
-
-  class ExtPackAST,
-
-  class InfiniteAST,
-
-  class VarTupleAST,
-
-  class ForwardAST,
-  class CheckEqAST,
-  class CheckIsinAST,
-  class FromToAST,
-
-  class FmtStrAST,
-
-  class ResourceAST,
-
-  class LocalPathAST,
-  class RemotePathAST,
-  class WebUrlAST,
-  class DBUrlAST,
-
-  class PrintAST,
-  class ReadFileAST>;
-
-/* Styio -> LLVM Generator */
-class StyioToLLVM;
-
-class StyioToLLVM : public StyioVisitor
-{
-  unique_ptr<llvm::LLVMContext> theContext;
-  unique_ptr<llvm::Module> theModule;
-  unique_ptr<llvm::IRBuilder<>> theBuilder;
-
-  std::unique_ptr<StyioJIT_ORC> theORCJIT;
-
-  unique_ptr<llvm::FunctionPassManager> theFPM;
-  unique_ptr<llvm::LoopAnalysisManager> theLAM;
-  unique_ptr<llvm::FunctionAnalysisManager> theFAM;
-  unique_ptr<llvm::CGSCCAnalysisManager> theCGAM;
-  unique_ptr<llvm::ModuleAnalysisManager> theMAM;
-  unique_ptr<llvm::PassInstrumentationCallbacks> thePIC;
-  unique_ptr<llvm::StandardInstrumentations> theSI;
-  llvm::PassBuilder thePB;
-
-  unordered_map<string, llvm::AllocaInst*> mut_vars; /* [FlexBind] Mutable Variables */
-  unordered_map<string, llvm::Value*> named_values;  /* [FinalBind] Named Values = Immutable Variables */
-
-  unordered_map<string, FuncAST*> func_defs;
-
-public:
-  StyioToLLVM(std::unique_ptr<StyioJIT_ORC> styio_jit) :
-      theContext(std::make_unique<llvm::LLVMContext>()),
-      theModule(std::make_unique<llvm::Module>("styio", *theContext)),
-      theBuilder(std::make_unique<llvm::IRBuilder<>>(*theContext)),
-      theORCJIT(std::move(styio_jit)),
-      theFPM(std::make_unique<llvm::FunctionPassManager>()),
-      theLAM(std::make_unique<llvm::LoopAnalysisManager>()),
-      theFAM(std::make_unique<llvm::FunctionAnalysisManager>()),
-      theCGAM(std::make_unique<llvm::CGSCCAnalysisManager>()),
-      theMAM(std::make_unique<llvm::ModuleAnalysisManager>()),
-      thePIC(std::make_unique<llvm::PassInstrumentationCallbacks>()),
-      theSI(std::make_unique<llvm::StandardInstrumentations>(*theContext, /*DebugLogging*/ true)) {
-    theModule->setDataLayout(theORCJIT->getDataLayout());
-
-    theSI->registerCallbacks(*thePIC, theMAM.get());
-
-    // Add transform passes.
-    // Do simple "peephole" optimizations and bit-twiddling optimizations.
-    theFPM->addPass(llvm::InstCombinePass());
-    // Reassociate expressions.
-    theFPM->addPass(llvm::ReassociatePass());
-    // Eliminate common sub-expressions.
-    theFPM->addPass(llvm::GVNPass());
-    // Simplify the control flow graph (deleting unreachable blocks, etc).
-    theFPM->addPass(llvm::SimplifyCFGPass());
-
-    thePB.registerModuleAnalyses(*theMAM);
-    thePB.registerFunctionAnalyses(*theFAM);
-    thePB.crossRegisterProxies(*theLAM, *theFAM, *theCGAM, *theMAM);
-  }
-
-  ~StyioToLLVM() {}
-
-  /* Utility Functions */
-  llvm::Type* matchType(string type);
-
-  llvm::AllocaInst* createAllocaFuncEntry(llvm::Function* TheFunction, llvm::StringRef VarName);
-
-  /* Styio AST Type typeInferer*/
-
-  void typeInfer(BoolAST* ast);
-
-  void typeInfer(NoneAST* ast);
-
-  void typeInfer(EOFAST* ast);
-
-  void typeInfer(EmptyAST* ast);
-
-  void typeInfer(PassAST* ast);
-
-  void typeInfer(BreakAST* ast);
-
-  void typeInfer(ReturnAST* ast);
-
-  void typeInfer(CommentAST* ast);
-
-  void typeInfer(IdAST* ast);
-
-  void typeInfer(VarAST* ast);
-
-  void typeInfer(ArgAST* ast);
-
-  void typeInfer(OptArgAST* ast);
-
-  void typeInfer(OptKwArgAST* ast);
-
-  void typeInfer(VarTupleAST* ast);
-
-  void typeInfer(DTypeAST* ast);
-
-  void typeInfer(IntAST* ast);
-
-  void typeInfer(FloatAST* ast);
-
-  void typeInfer(CharAST* ast);
-
-  void typeInfer(StringAST* ast);
-
-  void typeInfer(TypeConvertAST* ast);
-
-  void typeInfer(FmtStrAST* ast);
-
-  void typeInfer(LocalPathAST* ast);
-
-  void typeInfer(RemotePathAST* ast);
-
-  void typeInfer(WebUrlAST* ast);
-
-  void typeInfer(DBUrlAST* ast);
-
-  void typeInfer(ListAST* ast);
-
-  void typeInfer(TupleAST* ast);
-
-  void typeInfer(SetAST* ast);
-
-  void typeInfer(RangeAST* ast);
-
-  void typeInfer(SizeOfAST* ast);
-
-  void typeInfer(BinOpAST* ast);
-
-  void typeInfer(BinCompAST* ast);
-
-  void typeInfer(CondAST* ast);
-
-  void typeInfer(CallAST* ast);
-
-  void typeInfer(ListOpAST* ast);
-
-  void typeInfer(ResourceAST* ast);
-
-  void typeInfer(FlexBindAST* ast);
-
-  void typeInfer(FinalBindAST* ast);
-
-  void typeInfer(StructAST* ast);
-
-  void typeInfer(ReadFileAST* ast);
-
-  void typeInfer(PrintAST* ast);
-
-  void typeInfer(ExtPackAST* ast);
-
-  void typeInfer(BlockAST* ast);
-
-  void typeInfer(CasesAST* ast);
-
-  void typeInfer(CondFlowAST* ast);
-
-  void typeInfer(CheckEqAST* ast);
-
-  void typeInfer(CheckIsinAST* ast);
-
-  void typeInfer(FromToAST* ast);
-
-  void typeInfer(ForwardAST* ast);
-
-  void typeInfer(InfiniteAST* ast);
-
-  void typeInfer(AnonyFuncAST* ast);
-
-  void typeInfer(FuncAST* ast);
-
-  void typeInfer(LoopAST* ast);
-
-  void typeInfer(IterAST* ast);
-
-  void typeInfer(MatchCasesAST* ast);
-
-  void typeInfer(MainBlockAST* ast);
-
-  /* Get LLVM Type */
-
-  llvm::Type* getLLVMType(BoolAST* ast);
-
-  llvm::Type* getLLVMType(NoneAST* ast);
-
-  llvm::Type* getLLVMType(EOFAST* ast);
-
-  llvm::Type* getLLVMType(EmptyAST* ast);
-
-  llvm::Type* getLLVMType(PassAST* ast);
-
-  llvm::Type* getLLVMType(BreakAST* ast);
-
-  llvm::Type* getLLVMType(ReturnAST* ast);
-
-  llvm::Type* getLLVMType(CommentAST* ast);
-
-  llvm::Type* getLLVMType(IdAST* ast);
-
-  llvm::Type* getLLVMType(VarAST* ast);
-
-  llvm::Type* getLLVMType(ArgAST* ast);
-
-  llvm::Type* getLLVMType(OptArgAST* ast);
-
-  llvm::Type* getLLVMType(OptKwArgAST* ast);
-
-  llvm::Type* getLLVMType(VarTupleAST* ast);
-
-  llvm::Type* getLLVMType(DTypeAST* ast);
-
-  llvm::Type* getLLVMType(IntAST* ast);
-
-  llvm::Type* getLLVMType(FloatAST* ast);
-
-  llvm::Type* getLLVMType(CharAST* ast);
-
-  llvm::Type* getLLVMType(StringAST* ast);
-
-  llvm::Type* getLLVMType(TypeConvertAST* ast);
-
-  llvm::Type* getLLVMType(FmtStrAST* ast);
-
-  llvm::Type* getLLVMType(ListAST* ast);
-
-  llvm::Type* getLLVMType(TupleAST* ast);
-
-  llvm::Type* getLLVMType(SetAST* ast);
-
-  llvm::Type* getLLVMType(RangeAST* ast);
-
-  llvm::Type* getLLVMType(SizeOfAST* ast);
-
-  llvm::Type* getLLVMType(BinOpAST* ast);
-
-  llvm::Type* getLLVMType(BinCompAST* ast);
-
-  llvm::Type* getLLVMType(CondAST* ast);
-
-  llvm::Type* getLLVMType(CallAST* ast);
-
-  llvm::Type* getLLVMType(ListOpAST* ast);
-
-  llvm::Type* getLLVMType(ResourceAST* ast);
-
-  llvm::Type* getLLVMType(LocalPathAST* ast);
-
-  llvm::Type* getLLVMType(RemotePathAST* ast);
-
-  llvm::Type* getLLVMType(WebUrlAST* ast);
-
-  llvm::Type* getLLVMType(DBUrlAST* ast);
-
-  llvm::Type* getLLVMType(FlexBindAST* ast);
-
-  llvm::Type* getLLVMType(FinalBindAST* ast);
-
-  llvm::Type* getLLVMType(StructAST* ast);
-
-  llvm::Type* getLLVMType(ReadFileAST* ast);
-
-  llvm::Type* getLLVMType(PrintAST* ast);
-
-  llvm::Type* getLLVMType(ExtPackAST* ast);
-
-  llvm::Type* getLLVMType(BlockAST* ast);
-
-  llvm::Type* getLLVMType(CasesAST* ast);
-
-  llvm::Type* getLLVMType(CondFlowAST* ast);
-
-  llvm::Type* getLLVMType(CheckEqAST* ast);
-
-  llvm::Type* getLLVMType(CheckIsinAST* ast);
-
-  llvm::Type* getLLVMType(FromToAST* ast);
-
-  llvm::Type* getLLVMType(ForwardAST* ast);
-
-  llvm::Type* getLLVMType(InfiniteAST* ast);
-
-  llvm::Type* getLLVMType(AnonyFuncAST* ast);
-
-  llvm::Type* getLLVMType(FuncAST* ast);
-
-  llvm::Type* getLLVMType(LoopAST* ast);
-
-  llvm::Type* getLLVMType(IterAST* ast);
-
-  llvm::Type* getLLVMType(MatchCasesAST* ast);
-
-  llvm::Type* getLLVMType(MainBlockAST* ast);
-
-  /* LLVM IR Generator */
-
-  llvm::Value* toLLVMIR(BoolAST* ast);
-
-  llvm::Value* toLLVMIR(NoneAST* ast);
-
-  llvm::Value* toLLVMIR(EOFAST* ast);
-
-  llvm::Value* toLLVMIR(EmptyAST* ast);
-
-  llvm::Value* toLLVMIR(PassAST* ast);
-
-  llvm::Value* toLLVMIR(BreakAST* ast);
-
-  llvm::Value* toLLVMIR(ReturnAST* ast);
-
-  llvm::Value* toLLVMIR(CommentAST* ast);
-
-  llvm::Value* toLLVMIR(IdAST* ast);
-
-  llvm::Value* toLLVMIR(VarAST* ast);
-
-  llvm::Value* toLLVMIR(ArgAST* ast);
-
-  llvm::Value* toLLVMIR(OptArgAST* ast);
-
-  llvm::Value* toLLVMIR(OptKwArgAST* ast);
-
-  llvm::Value* toLLVMIR(VarTupleAST* ast);
-
-  llvm::Value* toLLVMIR(DTypeAST* ast);
-
-  llvm::Value* toLLVMIR(IntAST* ast);
-
-  llvm::Value* toLLVMIR(FloatAST* ast);
-
-  llvm::Value* toLLVMIR(CharAST* ast);
-
-  llvm::Value* toLLVMIR(StringAST* ast);
-
-  llvm::Value* toLLVMIR(TypeConvertAST* ast);
-
-  llvm::Value* toLLVMIR(FmtStrAST* ast);
-
-  llvm::Value* toLLVMIR(ListAST* ast);
-
-  llvm::Value* toLLVMIR(TupleAST* ast);
-
-  llvm::Value* toLLVMIR(SetAST* ast);
-
-  llvm::Value* toLLVMIR(RangeAST* ast);
-
-  llvm::Value* toLLVMIR(SizeOfAST* ast);
-
-  llvm::Value* toLLVMIR(BinOpAST* ast);
-
-  llvm::Value* toLLVMIR(BinCompAST* ast);
-
-  llvm::Value* toLLVMIR(CondAST* ast);
-
-  llvm::Value* toLLVMIR(CallAST* ast);
-
-  llvm::Value* toLLVMIR(ListOpAST* ast);
-
-  llvm::Value* toLLVMIR(ResourceAST* ast);
-
-  llvm::Value* toLLVMIR(LocalPathAST* ast);
-
-  llvm::Value* toLLVMIR(RemotePathAST* ast);
-
-  llvm::Value* toLLVMIR(WebUrlAST* ast);
-
-  llvm::Value* toLLVMIR(DBUrlAST* ast);
-
-  llvm::Value* toLLVMIR(FlexBindAST* ast);
-
-  llvm::Value* toLLVMIR(FinalBindAST* ast);
-
-  llvm::Value* toLLVMIR(StructAST* ast);
-
-  llvm::Value* toLLVMIR(ReadFileAST* ast);
-
-  llvm::Value* toLLVMIR(PrintAST* ast);
-
-  llvm::Value* toLLVMIR(ExtPackAST* ast);
-
-  llvm::Value* toLLVMIR(BlockAST* ast);
-
-  llvm::Value* toLLVMIR(CasesAST* ast);
-
-  llvm::Value* toLLVMIR(CondFlowAST* ast);
-
-  llvm::Value* toLLVMIR(CheckEqAST* ast);
-
-  llvm::Value* toLLVMIR(CheckIsinAST* ast);
-
-  llvm::Value* toLLVMIR(FromToAST* ast);
-
-  llvm::Value* toLLVMIR(ForwardAST* ast);
-
-  llvm::Value* toLLVMIR(InfiniteAST* ast);
-
-  llvm::Value* toLLVMIR(AnonyFuncAST* ast);
-
-  llvm::Value* toLLVMIR(FuncAST* ast);
-
-  llvm::Value* toLLVMIR(LoopAST* ast);
-
-  llvm::Value* toLLVMIR(IterAST* ast);
-
-  llvm::Value* toLLVMIR(MatchCasesAST* ast);
-
-  // llvm::Value* toLLVMIR(MainBlockAST* ast);
-  llvm::Function* toLLVMIR(MainBlockAST* ast);
-
-  void print_type_infer(StyioAST* program);
-
-  void print_llvm_ir();
-  void print_test_results();
-
-  void execute();
-};
+/* ========================================================================== */
 
 /*
   StyioAST: Styio Base AST
@@ -797,14 +36,16 @@ public:
   ) = 0;
 
   /* Type Inference */
-  virtual void typeInfer(StyioToLLVM* visitor) = 0;
+  virtual void typeInfer(StyioASTAnalyzer* visitor) = 0;
 
   /* Get LLVM Type */
-  virtual llvm::Type* getLLVMType(StyioToLLVM* generator) = 0;
+  virtual llvm::Type* getLLVMType(StyioToLLVMIR* visitor) = 0;
 
   /* LLVM IR Generator */
-  virtual llvm::Value* toLLVMIR(StyioToLLVM* generator) = 0;
+  virtual llvm::Value* toLLVMIR(StyioToLLVMIR* visitor) = 0;
 };
+
+/* ========================================================================== */
 
 template <class Derived>
 class StyioNode : public StyioAST
@@ -814,19 +55,20 @@ public:
   using StyioAST::toString;
   using StyioAST::toStringInline;
 
-  void typeInfer(StyioToLLVM* visitor) override {
+  void typeInfer(StyioASTAnalyzer* visitor) override {
     visitor->typeInfer(static_cast<Derived*>(this));
   }
 
-  llvm::Type* getLLVMType(StyioToLLVM* visitor) override {
+  llvm::Type* getLLVMType(StyioToLLVMIR* visitor) override {
     return visitor->getLLVMType(static_cast<Derived*>(this));
   }
 
-  llvm::Value* toLLVMIR(StyioToLLVM* visitor) override {
-    visitor->typeInfer(static_cast<Derived*>(this));
+  llvm::Value* toLLVMIR(StyioToLLVMIR* visitor) override {
     return visitor->toLLVMIR(static_cast<Derived*>(this));
   }
 };
+
+/* ========================================================================== */
 
 class CommentAST : public StyioNode<CommentAST>
 {
