+++ src/include/StyioUtil/Util.hpp
@@ -24,10 +24,22 @@ show_vector(std::vector<T> v) {
 
 template <typename T>
 void
-print_ast(T& program) {
-  std::cout << "\033[1;32mAST\033[0m \033[31m--No-Type-Checking\033[0m" << "\n" << std::endl;
-  std::cout << program->toString() << std::endl;
-  std::cout << "\n" << std::endl;
+print_ast(T& program, bool after_check) {
+  if (after_check) {
+    std::cout 
+      << "\033[1;32mAST\033[0m \033[31m-No-Type-Checking\033[0m" 
+      << "\n" 
+      << program->toString() 
+      << "\n" << std::endl;
+  }
+  else {
+    std::cout 
+      << "\033[1;32mAST\033[0m \033[1;33m-After-Type-Checking\033[0m"
+      << "\n"
+      << program->toString() 
+      << "\n"
+      << std::endl;
+  }
 };
 
 inline bool
